- added the possibility to know the OS specific thread id of a node - getEmitter and getCollector moved from protected to public

+++ ff/mapping_utils.hpp
@@ -79,7 +79,7 @@
  *  negative vlue is returned.
  *
  */
-static inline long ff_getThreadID() {
+static inline size_t ff_getThreadID() {
 #if (defined(__GNUC__) && defined(__linux))
     return  gettid();
 #elif defined(__APPLE__) && MAC_OS_X_HAS_AFFINITY
